Tweak the graph structure so that it uses the right equivalence notion.
Generic hash is forbidden on Coq structures, as they internally rely on
data that break its expected properties. Instead of relying on the
generic hash table structure, we use an ordered map. This also ensures
that the output is ordered w.r.t. the standard order on references,
which makes it more robust.